圧力センサーの値が10N単位から1N単位になったのに対応(実機と同じ).

+++ src.webots/jp/ac/fit/asura/nao/webots/Webots6Driver.java
@@ -150,8 +150,8 @@ class Webots6Driver {
 			return (float) ((gyro.getValues())[0]);
 		}
 
-		public int getForce(PressureSensor ts) {
-			return (int) fsr.get(ts).getValue();
+		public float getForce(PressureSensor ts) {
+			return (float) fsr.get(ts).getValue();
 		}
 
 		public float getForce(Joint joint) {
